query param items w/ base64 encoding
but base64 SUX for what we want to do. We controll the data, so we should come up with a more concise way to encode things. E.g., instead of 'helm' we can have '1'; there are just 12 item slots (15-16 if we count all others, but we can add them later..); same we can give a numeric ID to all items instead of using their full name. The questio then becomes 'how do we keep the name-id pairs consistent if we need to re-scrape the items'